refactor(renderer): split ConversationView into stores and components
ConversationView.tsx carried 4077 lines behind a 200-field scope object.
State now lives in one createStore/factory per concern under
conversation/stores (composer, queue, activity, browser, voice, search,
scroll, panels, settings, skills, message actions), composed by
conversation-scope.ts; Header/Timeline/Composer/Panels/Overlays are
separate files reading the same scope. No visual or contract changes.
