+++ src/renderer/src/components/conversation/composer-draft.ts
@@ -1,0 +1,15 @@
+import type { ComposerDraft } from "../ConversationView";
+
+export const EMPTY_DRAFT: ComposerDraft = {
+  text: "",
+  attachments: [],
+  replyToMessageId: null,
+};
+
+export function copyComposerDraft(draft: ComposerDraft): ComposerDraft {
+  return {
+    text: draft.text,
+    attachments: [...draft.attachments],
+    replyToMessageId: draft.replyToMessageId,
+  };
+}
