+++ src/renderer/src/components/conversation/stores/queue-store.ts
@@ -1,0 +1,57 @@
+import type { QueueDelivery } from "@openbot/contracts/ipc";
+import { createEffect, createMemo, createSignal, onCleanup, untrack } from "solid-js";
+import { activeQueueDeliveries, presentQueueDeliveries, queuedDeliveriesInOrder } from "../../../queue-reconciliation";
+import type { ConversationProps } from "../../ConversationView";
+import { agentActivityExitDuration } from "../activity-timing";
+
+export interface QueueStoreDeps {
+  props: ConversationProps;
+}
+
+export function createQueueStore(deps: QueueStoreDeps) {
+  const activeDeliveries = createMemo(() => activeQueueDeliveries(deps.props.queue, deps.props.activeTurnId));
+  const orderedQueuedDeliveries = createMemo(() => queuedDeliveriesInOrder(deps.props.queue));
+  const presentedQueueDeliveries = createMemo(() =>
+    presentQueueDeliveries({
+      snapshot: deps.props.queue,
+      activeTurnId: deps.props.activeTurnId,
+      renderedMessageIds: new Set(deps.props.messages.map((message) => message.id)),
+    }),
+  );
+  const [renderedQueueDeliveries, setRenderedQueueDeliveries] = createSignal<QueueDelivery[]>([]);
+  const queuePanelVisible = createMemo(() => renderedQueueDeliveries().length > 0);
+  let queueExitTimer: number | undefined;
+  createEffect(
+    () => presentedQueueDeliveries(),
+    (deliveries) => {
+      if (queueExitTimer !== undefined) {
+        window.clearTimeout(queueExitTimer);
+        queueExitTimer = undefined;
+      }
+      if (deliveries.length > 0) {
+        setRenderedQueueDeliveries(deliveries);
+        return;
+      }
+      if (untrack(renderedQueueDeliveries).length === 0) return;
+      queueExitTimer = window.setTimeout(() => {
+        queueExitTimer = undefined;
+        if (untrack(presentedQueueDeliveries).length === 0) setRenderedQueueDeliveries([]);
+      }, agentActivityExitDuration());
+    },
+  );
+  onCleanup(() => {
+    if (queueExitTimer !== undefined) window.clearTimeout(queueExitTimer);
+  });
+
+  return {
+    activeDeliveries,
+    orderedQueuedDeliveries,
+    presentedQueueDeliveries,
+    renderedQueueDeliveries,
+    setRenderedQueueDeliveries,
+    queuePanelVisible,
+    getQueueExitTimer: () => queueExitTimer,
+  };
+}
+
+export type QueueStore = ReturnType<typeof createQueueStore>;
