+++ src/renderer/src/conversation-keys.test.ts
@@ -1,0 +1,17 @@
+import { describe, expect, it } from "vitest";
+import { agentConversationKey, agentMessageKey, composerDraftKey } from "./conversation-keys";
+
+describe("conversation keys", () => {
+  it("keys drafts by server and bot", () => {
+    expect(composerDraftKey({ botId: "chief", serverId: "local" })).toBe("chief");
+    expect(composerDraftKey({ botId: "chief", serverId: "team-1" })).toBe("team-1:chief");
+    expect(composerDraftKey({ botId: "chief", serverId: "team-1" })).not.toBe(
+      composerDraftKey({ botId: "chief", serverId: "team-2" }),
+    );
+  });
+
+  it("keeps conversation and message keys unambiguous", () => {
+    expect(agentConversationKey("s", "a:b")).not.toBe(agentConversationKey("s:a", "b"));
+    expect(agentMessageKey("chief", "m1")).toBe("chief\0m1");
+  });
+});
